+++ lib/rundoc/code_command/background/stdin_write.rb
@@ -6,10 +6,15 @@ class Rundoc::CodeCommand::Background
     def initialize(contents, name:, wait:, timeout: 5, ending: $/)
       @contents = contents
       @ending = ending
-      @spawn = Rundoc::CodeCommand::Background::ProcessSpawn.find(name)
       @wait = wait
+      @name = name
       @timeout_value = Integer(timeout)
       @contents_written = nil
+      @background = nil
+    end
+
+    def background
+      @background ||= Rundoc::CodeCommand::Background::ProcessSpawn.find(@name)
     end
 
     # The command is rendered (`:::>-`) by the output of the `def call` method.
@@ -20,11 +25,11 @@ class Rundoc::CodeCommand::Background
     # The contents produced by the command (`:::->`) are rendered by the `def to_md` method.
     def call(env = {})
       writecontents
-      @spawn.log.read
+      background.log.read
     end
 
     def writecontents
-      @contents_written ||= @spawn.stdin_write(
+      @contents_written ||= background.stdin_write(
         contents,
         wait: @wait,
         ending: @ending,