+++ lib/rundoc/code_command/background/wait.rb
@@ -1,9 +1,14 @@
 class Rundoc::CodeCommand::Background
   class Wait < Rundoc::CodeCommand
     def initialize(name:, wait:, timeout: 5)
-      @spawn = Rundoc::CodeCommand::Background::ProcessSpawn.find(name)
+      @name = name
       @wait = wait
       @timeout_value = Integer(timeout)
+      @background = nil
+    end
+
+    def background
+      @background ||= Rundoc::CodeCommand::Background::ProcessSpawn.find(name)
     end
 
     def to_md(env = {})
@@ -11,7 +16,7 @@ class Rundoc::CodeCommand::Background
     end
 
     def call(env = {})
-      @spawn.wait(@wait, @timeout_value)
+      background.wait(@wait, @timeout_value)
       ""
     end
   end